long to ing

+++ help-desk-api/Models/Dto/UserManagement/UserFilterDto.cs
@@ -4,7 +4,7 @@ namespace Models.Dto.UserManagement
     public class UserFilterDto : PagedInputDto
     {
         public int CompanyId { get; set; }
-        public List<long> RoleIds { get; set; } = new List<long>();
-        public List<long> DepartmentIds { get; set; } = new List<long>();
+        public List<int> RoleIds { get; set; } = new List<int>();
+        public List<int> DepartmentIds { get; set; } = new List<int>();
     }
 }
