+++ help-desk-api/Models/Entities/Issue/TicketLeadCustomerMapModel.cs
@@ -4,11 +4,11 @@ using System.ComponentModel.DataAnnotations.Schema;
 namespace Models.Entities.Issue
 {
     [Table("TicketLeadCustomerMap", Schema = "issue")]
-    public class TicketLeadCustomerMapModel : BaseEntity<long>
+    public class TicketLeadCustomerMapModel : BaseEntity<int>
     {
 
-        public long FKTicketId { get; set; }
-        public long FKLeadCustomerId { get; set; }
+        public int FKTicketId { get; set; }
+        public int FKLeadCustomerId { get; set; }
 
         [ForeignKey("FKTicketId")]
         public TicketModel Ticket { get; set; }
