+++ help-desk-api/Models/Entities/UserManagement/MenuActionRoleMappingModel.cs
@@ -4,12 +4,12 @@ using Models.Entities;
 namespace Models.Entities.UserManagement
 {
     [Table("MenuActionRoleMapping", Schema = "UserMgmt")]
-    public class MenuActionRoleMappingModel : BaseEntity<long>
+    public class MenuActionRoleMappingModel : BaseEntity<int>
     {
         public int FKRoleId { get; set; }
         [ForeignKey("FKRoleId")]
         public RoleModel Role { get; set; }
-        public long FKMenuActionMapId { get; set; }
+        public int FKMenuActionMapId { get; set; }
         [ForeignKey("FKMenuActionMapId")]
         public MenuActionMapModel MenuActionMap { get; set; }
         public bool IsAllowed { get; set; }
